chore(mapbox): set the api token at build time
Since the Mapbox api token is declared as NEXT_PUBLIC..,
Nextjs inlines the value during build time. This is so
it's available to the client to use.

That makes it a bit more of a challenge when trying
to create a single immutable build for multiple
deployments. The solution captured here is to grab'
the token from the docker build environment so the
image can be built once.

This is the simpler scenario because the same key
is used in all environments. More work will have to
be done when adding environment values that are needed
during build time.